Edvantage/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.1.2
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
